HW12: AI for the game
Since the game is imperfect information game, the MAXMIN method is not
very practical here. So I customize my own rule about about the AI:
If it has the King, it will play the King.
If the AI has card which will gain advantage(i.e., 7 and J), it will
immediately played that card.
If the point is small than 50, it will play the largest card it has
without playing the rank Four card which is a very good wild card.
If the point is larger than 50, it will play the smallest.
If the smallest card is not appropriate, i.e. the total point will
still exceed 99, it will try to play the Rank Four card, which is the
last card it can play.
Otherwise, no matter what card it plays, it will lose that game.

The URL for the HW in emulator
http://5-dot-ninety-nine.appspot.com/